[BACKLOG-5807] Renamed chart directories to a more logical name
