+++ package-res/resources/web/pentaho/visual/ccc/abstract/i18n/model.properties
@@ -80,9 +80,9 @@ barStacked.props.columns.label=Color Stack
 barStacked.props.measures.label=Measures
 
 ## Stacked Bar Chart (Horizontal)
-barHorizontalStacked.props.rows.label=Y-Axis
-barHorizontalStacked.props.columns.label=Color Stack
-barHorizontalStacked.props.measures.label=Measures
+barStackedHorizontal.props.rows.label=Y-Axis
+barStackedHorizontal.props.columns.label=Color Stack
+barStackedHorizontal.props.measures.label=Measures
 
 ## Normalized Bar Chart (Vertical, Stacked)
 barNormalized.props.rows.label=X-Axis
@@ -90,9 +90,9 @@ barNormalized.props.columns.label=Color Stack
 barNormalized.props.measures.label=Measures
 
 ## Normalized Bar Chart (Horizontal, Stacked)
-barHorizontalNormalized.props.rows.label=X-Axis
-barHorizontalNormalized.props.columns.label=Color Stack
-barHorizontalNormalized.props.measures.label=Measures
+barNormalizedHorizontal.props.rows.label=X-Axis
+barNormalizedHorizontal.props.columns.label=Color Stack
+barNormalizedHorizontal.props.measures.label=Measures
 
 ## Sunburst
 sunburst.props.displayUnits.label=Scale
