+++ package-res/resources/web/pentaho/visual/ccc/barNormalizedHorizontal/View.js
