+++ package-res/resources/web/pentaho/visual/ccc/barStackedHorizontal/View.js
